# ai-model-validation-platform/backend/docs/openapi-multi-video.yaml
openapi: 3.0.0
info:
  title: Multi-Video HIL Testing API
  version: 1.0.0
  description: |
    API endpoints for Hardware-in-the-Loop (HIL) testing with multi-video sequence support.

    ## Multi-Video Architecture

    The system supports sequential playback of multiple videos in a single test session:

    1. **Test Session Creation**: Create session with multiple video IDs
    2. **Sequence Tracking**: Track video start/end times with nanosecond precision
    3. **Per-Video Results**: Store detection counts, latency, and pass rates per video
    4. **Video Filtering**: Filter results by specific video_id in multi-video sessions

    ## Database Schema

    - **TestSession**: has_video_sequence, sequence_id, sequence_metadata fields
    - **VideoTestSequence**: Container for multi-video test sequences
    - **SequenceVideoResult**: Per-video results within sequences
    - **DetectionEvent**: Linked to specific video via video_id and sequence_video_result_id

  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/video-sequences/{sequence_id}/video-started:
    post:
      summary: Record Video Start Event
      description: |
        Track when a video in a sequence starts playing.

        **Purpose**:
        - Record precise video start timestamp with nanosecond precision
        - Update SequenceVideoResult.video_start_time
        - Update VideoTestSequence.current_video_index
        - Emit WebSocket event for real-time frontend updates

        **Called by**: SequentialVideoPlayer component when video playback begins
      tags:
        - Video Sequences
      parameters:
        - name: sequence_id
          in: path
          required: true
          description: Video sequence identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoStartedRequest'
            example:
              videoId: "video-002"
              startedAt: 1730307890.123456
              sequenceElapsedTime: 30.5
              timestamp: 1730307890.123456
      responses:
        '200':
          description: Video start event recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoEventResponse'
              example:
                status: "acknowledged"
                message: "Video video-002 start recorded"
        '404':
          description: Sequence or video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Video sequence seq-abc123 not found"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/video-sequences/{sequence_id}/video-ended:
    post:
      summary: Record Video End Event
      description: |
        Track when a video in a sequence ends playing.

        **Purpose**:
        - Record precise video end timestamp with nanosecond precision
        - Update SequenceVideoResult.video_end_time and actual_duration_ms
        - Increment VideoTestSequence.completed_videos
        - Return next video ID if available
        - Emit WebSocket event for real-time frontend updates

        **Called by**: SequentialVideoPlayer component when video playback completes
      tags:
        - Video Sequences
      parameters:
        - name: sequence_id
          in: path
          required: true
          description: Video sequence identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoEndedRequest'
            example:
              videoId: "video-002"
              endedAt: 1730307920.456789
              actualDuration: 25.5
              sequenceElapsedTime: 56.0
              timestamp: 1730307920.456789
      responses:
        '200':
          description: Video end event recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoEventResponse'
              examples:
                with_next_video:
                  summary: More videos remaining in sequence
                  value:
                    status: "acknowledged"
                    nextVideoId: "video-003"
                    message: "Video video-002 completion recorded"
                sequence_complete:
                  summary: Last video in sequence
                  value:
                    status: "acknowledged"
                    nextVideoId: null
                    message: "Video video-003 completion recorded"
        '404':
          description: Sequence or video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/video-sequences/{sequence_id}/status:
    get:
      summary: Get Sequence Status
      description: |
        Retrieve current status and progress of a video sequence.

        **Returns**:
        - Overall sequence status (pending, playing, completed, archived)
        - Current video index and ID
        - Completion progress (videos completed / total videos)
        - Per-video results including detection counts and latency
      tags:
        - Video Sequences
      parameters:
        - name: sequence_id
          in: path
          required: true
          description: Video sequence identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sequence status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SequenceStatus'
              example:
                sequenceId: "seq-abc123"
                testSessionId: "session-789"
                status: "playing"
                startedAt: "2025-10-30T10:00:00Z"
                currentVideo: "video-002"
                currentVideoIndex: 1
                videosCompleted: 1
                totalVideos: 3
                videoIds: ["video-001", "video-002", "video-003"]
                videos:
                  - videoId: "video-001"
                    sequenceOrder: 0
                    status: "completed"
                    startedAt: 1730307890.0
                    endedAt: 1730307920.0
                    durationMs: 30000
                    detectionCount: 5
                    passedDetections: 4
                    failedDetections: 1
                    avgLatencyMs: 78.5
                    passRatePercent: 80.0
                    validationResult: "pass"
                  - videoId: "video-002"
                    sequenceOrder: 1
                    status: "playing"
                    startedAt: 1730307920.0
                    endedAt: null
                    durationMs: null
                    detectionCount: 0
                    passedDetections: 0
                    failedDetections: 0
                    avgLatencyMs: null
                    passRatePercent: null
                    validationResult: null
                  - videoId: "video-003"
                    sequenceOrder: 2
                    status: "pending"
                    startedAt: null
                    endedAt: null
                    durationMs: null
                    detectionCount: 0
                    passedDetections: 0
                    failedDetections: 0
                    avgLatencyMs: null
                    passRatePercent: null
                    validationResult: null
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    VideoStartedRequest:
      type: object
      required:
        - videoId
      properties:
        videoId:
          type: string
          format: uuid
          description: Video identifier
          example: "video-002"
        startedAt:
          type: number
          format: double
          description: Video start timestamp in seconds (preferred field)
          example: 1730307890.123456
        timestamp:
          type: number
          format: double
          description: Legacy timestamp field (seconds)
          example: 1730307890.123456
        sequenceElapsedTime:
          type: number
          format: double
          description: Elapsed time in sequence (seconds)
          default: 0.0
          example: 30.5
        clientTimestamp:
          type: string
          format: date-time
          nullable: true
          description: Optional ISO timestamp from frontend
    VideoEndedRequest:
      type: object
      required:
        - videoId
      properties:
        videoId:
          type: string
          format: uuid
          description: Video identifier
          example: "video-002"
        endedAt:
          type: number
          format: double
          description: Video end timestamp in seconds (preferred field)
          example: 1730307920.456789
        timestamp:
          type: number
          format: double
          description: Legacy timestamp field (seconds)
          example: 1730307920.456789
        actualDuration:
          type: number
          format: double
          description: Actual playback duration in seconds
          example: 25.5
        sequenceElapsedTime:
          type: number
          format: double
          description: Elapsed time in sequence (seconds)
          example: 56.0
        clientTimestamp:
          type: string
          format: date-time
          nullable: true
          description: Optional ISO timestamp from frontend
    VideoEventResponse:
      type: object
      properties:
        status:
          type: string
          description: Event processing status
          example: "acknowledged"
        nextVideoId:
          type: string
          format: uuid
          nullable: true
          description: Next video ID if available (null if sequence complete)
          example: "video-003"
        message:
          type: string
          nullable: true
          description: Response message
          example: "Video video-002 start recorded"
    SequenceStatus:
      type: object
      properties:
        sequenceId:
          type: string
          format: uuid
        testSessionId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, playing, completed, archived]
        startedAt:
          type: string
          format: date-time
          nullable: true
        currentVideo:
          type: string
          format: uuid
          nullable: true
        currentVideoIndex:
          type: integer
          nullable: true
        videosCompleted:
          type: integer
        totalVideos:
          type: integer
        videoIds:
          type: array
          items:
            type: string
        videos:
          type: array
          items:
            $ref: '#/components/schemas/VideoResult'
    VideoResult:
      type: object
      properties:
        videoId:
          type: string
          format: uuid
        sequenceOrder:
          type: integer
        status:
          type: string
          enum: [pending, playing, completed, failed]
        startedAt:
          type: number
          format: double
          nullable: true
        endedAt:
          type: number
          format: double
          nullable: true
        durationMs:
          type: number
          format: double
          nullable: true
        detectionCount:
          type: integer
        passedDetections:
          type: integer
        failedDetections:
          type: integer
        avgLatencyMs:
          type: number
          format: double
          nullable: true
        passRatePercent:
          type: number
          format: double
          nullable: true
        validationResult:
          type: string
          enum: [pass, fail, pending]
          nullable: true
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: "Resource not found"
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Video sequence seq-abc123 not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Failed to process request: database connection lost"
security: []